+++ core/security/jwt/blolet-jwt-filter/src/main/java/nettee/blolet/jwt/filter/JwtAuthorizationFilter.java
@@ -1,4 +1,4 @@
-package nettee.jwt.filter;
+package nettee.blolet.jwt.filter;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
