+++ core/security/jwt/blolet-jwt-filter/src/main/java/nettee/blolet/jwt/filter/JwtFilterParserConfig.java
@@ -1,6 +1,7 @@
-package nettee.jwt.filter;
+package nettee.blolet.jwt.filter;
 
 import lombok.RequiredArgsConstructor;
+import nettee.blolet.jwt.filter.properties.AuthorizationFilterProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.util.pattern.PathPatternParser;
