+++ core/security/jwt/blolet-jwt-filter/src/main/java/nettee/blolet/jwt/filter/annotation/AuthorizedUser.java
@@ -1,4 +1,4 @@
-package nettee.jwt.annotation;
+package nettee.blolet.jwt.filter.annotation;
 
 import java.util.List;
