+++ core/security/jwt/blolet-jwt-filter/src/main/java/nettee/blolet/jwt/filter/exception/PathFilterErrorCode.java
@@ -1,4 +1,4 @@
-package nettee.jwt.filter;
+package nettee.blolet.jwt.filter.exception;
 
 import nettee.common.ErrorCode;
 import org.springframework.http.HttpStatus;
