+++ core/security/jwt/blolet-jwt-filter/src/main/java/nettee/blolet/jwt/filter/exception/PathFilterException.java
@@ -1,4 +1,4 @@
-package nettee.jwt.filter;
+package nettee.blolet.jwt.filter.exception;
 
 import nettee.common.CustomException;
 import nettee.common.ErrorCode;
