+++ services/auth/driving/web-mvc/src/main/java/nettee/auth/web/AuthCommandApi.java
@@ -15,8 +15,8 @@ import nettee.auth.web.dto.AuthCommandDto.SignUpRequest;
 import nettee.auth.web.mapper.AuthDtoMapper;
 import nettee.blolet.auth.readmodel.AuthCommandModels.LoginTokenModel;
 import nettee.blolet.auth.readmodel.AuthCommandModels.SignUpRequestModel;
-import nettee.jwt.annotation.AuthUser;
-import nettee.jwt.annotation.AuthorizedUser;
+import nettee.blolet.jwt.filter.annotation.AuthUser;
+import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
