+++ services/blog/driving/web-mvc/src/main/java/nettee/blolet/blog/web/BlogCommandApi.java
@@ -12,8 +12,8 @@ import nettee.blolet.blog.web.dto.BlogCommandDto.BlogUnsubscribeResponse;
 import nettee.blolet.blog.web.dto.BlogCommandDto.BlogUpdateCommand;
 import nettee.blolet.blog.web.dto.BlogCommandDto.BlogUpdateResponse;
 import nettee.blolet.blog.web.mapper.BlogDtoMapper;
-import nettee.jwt.annotation.AuthUser;
-import nettee.jwt.annotation.AuthorizedUser;
+import nettee.blolet.jwt.filter.annotation.AuthUser;
+import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
