+++ services/blog/driving/web-mvc/src/main/java/nettee/blolet/blog/web/admin/BlogAdminCommandApi.java
@@ -7,8 +7,8 @@ import nettee.blolet.blog.web.admin.dto.BlogAdminCommandDto.BlogCreateCommand;
 import nettee.blolet.blog.web.admin.dto.BlogAdminCommandDto.BlogCreateResponse;
 import nettee.blolet.blog.web.admin.dto.BlogAdminCommandDto.BlogInternalCreateCommand;
 import nettee.blolet.blog.web.admin.mapper.BlogAdminDtoMapper;
-import nettee.jwt.annotation.AuthUser;
-import nettee.jwt.annotation.AuthorizedUser;
+import nettee.blolet.jwt.filter.annotation.AuthUser;
+import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
