+++ services/draft/driving/web-mvc/src/main/java/nettee/draft/driving/web/DraftCommandApi.java
@@ -12,8 +12,8 @@ import nettee.draft.driving.web.dto.DraftCommandDto.DraftCommandResponse;
 import nettee.draft.driving.web.dto.DraftCommandDto.DraftCreateCommand;
 import nettee.draft.driving.web.dto.DraftCommandDto.DraftUpdateCommand;
 import nettee.draft.driving.web.mapper.DraftDtoMapper;
-import nettee.jwt.annotation.AuthUser;
-import nettee.jwt.annotation.AuthorizedUser;
+import nettee.blolet.jwt.filter.annotation.AuthUser;
+import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
